wip: enumerator: Add macro to define an ad-hoc filtering enumerator
wip: That the macro includes the opening { is not so nice as editors
that track these (but don't evaluate macros) will mark this as syntax
error. I suppose we could add an ENUMERATOR_FILTER_END() macro that just
writes the closing } and then use {} within these two macros.

Another idea might be to add a venumerate method to the enumerator_t
interface and make that mandatory (basically move the code from enumerate
to venumerate) and provide a simple generic enumerate callback for all to
use that only contains a va_start, a call to venumerate and a va_end.
We'd also provide a VA_ARGS_VGET macro to get the arguments on the stack
from a va_list.
Then we could forward arguments from one enumerator to another (for example
for the cleaner enumerator - which is otherwise a pretty tough one, if
we want to avoid having to define a cleanup function and enumerator for
every use).